chore: improve settings grouping
