+++ apps/client/src/features/app-settings/panel/feature-settings-panel/UrlPresetsForm.tsx
@@ -15,7 +15,7 @@ import { handleLinks } from '../../../../common/utils/linkUtils';
 import { validateUrlPresetPath } from '../../../../common/utils/urlPresets';
 import * as Panel from '../PanelUtils';
 
-import style from './GeneralPanel.module.scss';
+import style from './FeatureSettings.module.scss';
 
 const urlPresetsDocs = 'https://docs.getontime.no/features/url-presets/';
