+++ apps/client/src/features/app-settings/panel/general-panel/GeneralPanel.module.scss
@@ -2,23 +2,3 @@
   display: flex;
   gap: 1em;
 }
-
-.pad {
-  padding: 0 2rem;
-}
-
-.fit {
-  width: fit-content;
-}
-
-.aliasConstrain {
-  min-width: 12em;
-}
-
-.fullWidth {
-  width: 100%;
-}
-
-.flex {
-  display: flex;
-}
