+++ apps/client/src/features/app-settings/panel/general-panel/GeneralPanel.tsx
@@ -1,27 +1,27 @@
 import useScrollIntoView from '../../../../common/hooks/useScrollIntoView';
 import { PanelBaseProps } from '../../settingsStore';
+import EditorSettingsForm from '../interface-panel/EditorSettingsForm';
 import * as Panel from '../PanelUtils';
 
 import GeneralPanelForm from './GeneralPanelForm';
-import UrlPresetsForm from './UrlPresetsForm';
 import ViewSettingsForm from './ViewSettingsForm';
 
 export default function GeneralPanel({ location }: PanelBaseProps) {
-  const manageRef = useScrollIntoView<HTMLDivElement>('manage', location);
+  const generalRef = useScrollIntoView<HTMLDivElement>('settings', location);
+  const editorRef = useScrollIntoView<HTMLDivElement>('editor', location);
   const viewRef = useScrollIntoView<HTMLDivElement>('view', location);
-  const urlPresetsRef = useScrollIntoView<HTMLDivElement>('urlpresets', location);
 
   return (
     <>
       <Panel.Header>App Settings</Panel.Header>
-      <div ref={manageRef}>
+      <div ref={generalRef}>
         <GeneralPanelForm />
+      </div>
+      <div ref={editorRef}>
+        <EditorSettingsForm />
       </div>
       <div ref={viewRef}>
         <ViewSettingsForm />
-      </div>
-      <div ref={urlPresetsRef}>
-        <UrlPresetsForm />
       </div>
     </>
   );
